# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0

paths:
  /tasks:
    get:
      summary: Get all tasks
      tags:
        - tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      tags:
        - tasks
      requestBody:
        description: The task to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: The created task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{task_id}:
    delete:
      summary: Delete a task by ID
      tags:
        - tasks
      operationId: DeleteTaskByID
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '204':
          description: Task Deleted
    patch:
      summary: Update a task by ID
      tags:
        - tasks
      operationId: UpdateTaskByID
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
      responses:
        '200':
          description: Task Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/usertasks/{user_id}:
    get:
      summary: Get all user tasks by ID
      tags:
        - tasks
      operationId: GetTasksByUserID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: A list of user tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /users:
    get:
      summary: Get all users
      tags:
        - users
      operationId: GetUsers
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: Create a new user
      tags:
        - users
      operationId: PostUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{user_id}:
    patch:
      summary: Update user data
      tags:
        - users
      operationId: PatchUserByID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      summary: Delete a user
      tags:
        - users
      operationId: DeleteUserByID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '204':
          description: User deleted

      
      

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: uint
        task:
          type: string
        is_done:
          type: boolean
        user_id:
          type: integer
          format: uint
    User:
      type: object
      properties:
        id:
          type: integer
          format: uint
        email:
          type: string
        password:
          type: string